# src/main/java/io/datastoria/server/dto/CreateAgentRevisionRequest.java
package io.datastoria.server.dto;

import jakarta.validation.constraints.NotBlank;

public record CreateAgentRevisionRequest(
    String modelId,
    @NotBlank String systemPrompt,
    String runtimeConfigJson,
    String toolPolicyJson,
    String skillPolicyJson) {}